helmcharts/camcom-sender/requirements.txt
boto3==1.28.63
cached_property==1.5.2
cachetools==5.3.3
cvat-sdk==2.8.0
dataclasses==0.6
dependency-injector==4.39.1
exif==1.2.1
Flask==2.1.0
flask-babel==3.0.1
Flask-Bootstrap==3.3.7.1
flask-cors==3.0.10
Werkzeug==2.3.6
geopy==2.2.0
gevent==21.12.0
gunicorn==20.0.4
haversine==2.8.1
httpx==0.24.1
kafka-python==2.0.2
opencv-python-headless==4.4.0.46
numpy==1.26.4
passlib[bcrypt]==1.7.4
piexif==1.1.3
Pillow==10.1.0
pre-commit==2.8.2
prometheus-flask-exporter==0.20.3
psycopg2-binary==2.9.9
pydantic==1.9.0
pyproj==3.2.1
pyquery==1.4.3
python-box==5.2.0
python-json-logger==2.0.4
requests==2.22.0
requests-ntlm==1.1.0
sentry-sdk[flask]==1.0.0
simplekml==1.3.5
sqlalchemy==1.4.7
geoalchemy2==0.14.4
ratelimiter==1.2.0.post0
utm==0.7.0
scikit-learn==1.0.2
fiona==1.9.6
shapely==2.0.1
yarl==1.9.4
h3==4.1.2
